refactor: split scoring internals into constants, tallies, and formulas
